+++ app/src/icons/Close.tsx
@@ -1,0 +1,24 @@
+import React, { ReactElement } from 'react';
+import { IconProps, useDefaultColor } from '.';
+import { DEFAULT_ICON_OPACITY, DEFAULT_ICON_SIZE } from '../../../common/const';
+
+export const Close = ({
+  color,
+  size = DEFAULT_ICON_SIZE,
+  opacity = DEFAULT_ICON_OPACITY,
+}: IconProps): ReactElement => {
+  const defaultColor = useDefaultColor();
+  const fill = color ?? defaultColor;
+
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      height={size}
+      width={size}
+      fill={fill}
+      opacity={opacity}
+    >
+      <path d="M6.062 15 5 13.938 8.938 10 5 6.062 6.062 5 10 8.938 13.938 5 15 6.062 11.062 10 15 13.938 13.938 15 10 11.062Z" />
+    </svg>
+  );
+};
