+++ common/const.ts
@@ -1,6 +1,8 @@
 import { Module } from './types/module';
 
 export const SIDE_BAR_WIDTH_PX = 56;
+export const DEFAULT_ICON_SIZE = 18;
+export const DEFAULT_ICON_OPACITY = 0.75;
 
 export const GOOGLE_FAVICON_URL = 'https://www.google.com/s2/favicons';
